+++ test-examples.sh
@@ -9,10 +9,10 @@ for EXAMPLE in examples/*; do
   echo "Testing $EXAMPLE ..."
   CMD="src/validate_submission.py \
 --local_directory $EXAMPLE/submission \
---dataset_ignore_globs ignore-me.tsv \
+--dataset_ignore_globs 'ignore-*.tsv' '.*' \
 --output as_md"
   README="$EXAMPLE/README.md"
-  diff $README <( $CMD ) \
+  diff $README <( eval "$CMD" ) \
     || die "Update example: $CMD > $README"
 done